Create separate script to log latlon plots to wandb
